requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
clickhouse-connect==0.8.17
clickhouse-driver==0.2.9
idna==3.10
lz4==4.4.4
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
tzlocal==5.3.1
urllib3==2.4.0
zstandard==0.23.0
